improve a bit the standalone lua example
